requirements.txt
asgiref                == 3.2.10
astroid                == 2.4.2
autopep8               == 1.5.4
brotlipy               == 0.7.0
# certifi                == 2020.6.20
cffi                   == 1.14.3
chardet                == 3.0.4
code-engine            == 1.0.13
codegenhelper          == 0.0.16
colorama               == 0.4.3
cryptography           == 3.1.1
cycler                 == 0.10.0
deep-mapper1           == 0.0.1
dj-database-url        ==  0.5.0
Django                 ==  3.1.1
django-cors-headers    ==  3.5.0
djangorestframework     == 3.11.1
djangorestframework-jwt == 1.11.0
dlib                    == 19.21.0
gunicorn               ==  20.0.4
idna                   ==  2.10
imutils                ==  0.5.3
isort                  ==  5.5.3
Jinja2                 ==  2.10
kiwisolver             ==  1.2.0
lazy-object-proxy      ==  1.4.3
mapper-on-file         ==  0.0.2
MarkupSafe             ==  1.1.1
matplotlib             ==  3.3.2
mccabe                 ==  0.6.1
# mkl-service            ==  2.3.0
nodejs                 ==  0.1.1
nodejs-codegen         ==  0.3.7
npm                    ==  0.1.1
# numpy                  ==  1.19.1
opencv-python           ==4.4.0.44
olefile                ==  0.46
optional-django        ==  0.1.0
Pillow                 ==  7.2.0
pip                    ==  20.2.2
psycopg2               ==  2.8.6
pycodestyle            ==  2.6.0
pycparser              ==  2.20
PyJWT                  ==  1.7.1
pylint                 ==  2.6.0
pyOpenSSL              ==  19.1.0
pyparsing              ==  2.4.7
PySocks                ==  1.7.1
python-dateutil        ==  2.8.1
pytz                   ==  2020.1
react                 ==   4.3.0
requests              ==   2.24.0
scipy                 ==   1.5.2
# setuptools            ==   49.6.0
six                   ==   1.15.0
sqlparse              ==   0.3.1
toml                  ==   0.10.1
tornado               ==   6.0.4
typed-ast              ==  1.4.1
urllib3                ==  1.25.10
websockets==8.1
wheel                  ==  0.35.1
whitenoise            ==  5.2.0
win-inet-pton          ==  1.1.0
wincertstore          ==   0.2
wrapt                 ==   1.11.2